+++ src/content-script-action/for-search-result-page.ts
@@ -3,7 +3,7 @@ import { ContentScriptAction } from "../content-script-action";
 import { GenreWordConverterFactory } from "../core/genre-word-converter-factory";
 import { GenreWordReplacerFactory } from "../core/genre-word-replacer-factory";
 import { GenreWordReplaceTargetPage } from "../core/genre-word-replace-target-page";
-import { DocumentTitleConverterFactory } from '../core/tab-title-converter-factory'
+import { DocumentTitleConverterFactory } from '../core/document-title-converter-factory'
 import { Message } from "../message";
 import { MessageType } from "../message/type";
 import {
