+++ src/core/document-title-converter-factory.ts
@@ -1,9 +1,9 @@
 
 import { GenreWordConversionMap } from "./genre-word-conversion-map";
 import { GenreWordConversionMode } from "./genre-word-conversion-mode";
-import { DocumentTitleConverter } from './tab-title-converter';
-import { DocumentTitleConverterToOldWords } from './tab-title-converter/to-old-words';
-import { DocumentTitleConverterToNewWords } from "./tab-title-converter/to-new-words";
+import { DocumentTitleConverter } from './document-title-converter';
+import { DocumentTitleConverterToOldWords } from './document-title-converter/to-old-words';
+import { DocumentTitleConverterToNewWords } from "./document-title-converter/to-new-words";
 
 export class DocumentTitleConverterFactory {
     public createDocumentTitleConverter(
